server/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
awsebcli==3.19.2
bcrypt==3.2.0
blessed==1.17.12
botocore==1.19.44
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
docker==4.4.1
docker-compose==1.25.5
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.20.0
future==0.16.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.1.1
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
packaging==20.8
paramiko==2.7.2
pathspec==0.5.9
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
semantic-version==2.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.4.0
zope.interface==5.2.0
